Rename file and export name of test-config file to config
