+++ QualityControl/test/public/index.js
@@ -16,7 +16,7 @@
 /* eslint-disable max-len */
 const puppeteer = require('puppeteer');
 const assert = require('assert');
-const config = require('../test-config.js');
+const { config } = require('../config.js');
 const { spawn } = require('child_process');
 
 /*
